refine DapAssetsPath import
