+++ src/lib/fabric-custom-shared.ts
@@ -1,21 +1,19 @@
 const CACHE_BUST = '%%CACHEBUSTER%%';
 const DapAssetsRoot = `https://s3-eu-west-1.amazonaws.com/adops-assets/dap-fabrics`;
 
-export const getDapAssetsPath = (DapAssetsFolder: string) =>
-	`${DapAssetsRoot}/${DapAssetsFolder}`;
-
-export const addTrackingPixel = (url: string) => {
+const addTrackingPixel = (url: string) => {
 	const pixel = new Image();
 	pixel.src = url + CACHE_BUST;
 };
 
 // relative paths in the CSS need to be replaced with the absolute path
-export const replaceAssetLinks = (html: string, DapAssetsPath: string) => {
+const replaceAssetLinks = (html: string, DapAssetsPath: string) => {
 	const re = /url\('\.\/(.*)'\)/g;
 	return html.replace(re, `url('${DapAssetsPath}/$1')`);
 };
 
-export const getTag = (DapAssetsFolder: string, DapAssetsPath: string) => {
+const getTag = (DapAssetsFolder: string) => {
+	const DapAssetsPath = `${DapAssetsRoot}/${DapAssetsFolder}`;
 	if (DapAssetsFolder) {
 		return fetch(`${DapAssetsPath}/index.html`)
 			.then((res) => res.text())
@@ -23,3 +21,13 @@ export const getTag = (DapAssetsFolder: string, DapAssetsPath: string) => {
 	}
 	return Promise.reject('No tag found');
 };
+
+const insertTag = (tag: string) => {
+	const placeholder = document.getElementById('creative-link')!;
+	const range = document.createRange();
+	range.setStart(placeholder, 0);
+	range.setEnd(placeholder, 0);
+	placeholder.appendChild(range.createContextualFragment(tag));
+};
+
+export { addTrackingPixel, replaceAssetLinks, getTag, insertTag };
